refactor requests to own folder and fix enum
